+++ bundles/admin/admin-layereditor/view/AdminLayerForm/GeneralTabPane/Name.jsx
@@ -1,0 +1,18 @@
+import React, { Fragment } from 'react';
+import PropTypes from 'prop-types';
+import { Message, TextInput } from 'oskari-ui';
+import { Controller } from 'oskari-ui/util';
+import { StyledComponent } from '../StyledFormComponents';
+
+export const Name = ({ layer, controller }) => (
+    <Fragment>
+        <Message messageKey='uniqueName' />
+        <StyledComponent>
+            <TextInput type='text' value={layer.name} onChange={evt => controller.setLayerName(evt.target.value)} />
+        </StyledComponent>
+    </Fragment>
+);
+Name.propTypes = {
+    layer: PropTypes.object.isRequired,
+    controller: PropTypes.instanceOf(Controller).isRequired
+};
